requirements.txt
playwright==1.44.0
pytest==9.0.2
pytest-playwright==0.5.0
pytest-html==4.1.1
pytest-asyncio==1.3.0
python-dotenv==1.0.1
allure-pytest==2.15.3
